+++ Core/Init.cs
@@ -2681,7 +2681,7 @@ namespace PortAIO
                             EloFactory_Riven.Program.Main();
                             break;
                         case 8: // Flower's Riven
-                            Flowers_Riven.Program.Main();
+                            Flowers_Riven_Reborn.Program.Main();
                             break;
                         case 9: // HeavenStrikeRiven
                             HeavenStrikeRiven.Program.Main();
