+++ PortAIO.csproj
@@ -1585,6 +1585,26 @@
     <Compile Include="Core\Champion Ports\Rengar\ElRengarDecentralized\Utils\ItemManager.cs" />
     <Compile Include="Core\Champion Ports\Rengar\ElRengarDecentralized\Utils\Logging.cs" />
     <Compile Include="Core\Champion Ports\Rengar\ElRengarDecentralized\Utils\PrioritizeManager.cs" />
+    <Compile Include="Core\Champion Ports\Riven\Flowers%27 Riven\Logic.cs" />
+    <Compile Include="Core\Champion Ports\Riven\Flowers%27 Riven\Manager\Events\Animation\AnimationManager.cs" />
+    <Compile Include="Core\Champion Ports\Riven\Flowers%27 Riven\Manager\Events\DoCast\DoCastManager.cs" />
+    <Compile Include="Core\Champion Ports\Riven\Flowers%27 Riven\Manager\Events\Drawings\DrawManager.cs" />
+    <Compile Include="Core\Champion Ports\Riven\Flowers%27 Riven\Manager\Events\EventManager.cs" />
+    <Compile Include="Core\Champion Ports\Riven\Flowers%27 Riven\Manager\Events\Games\LoopManager.cs" />
+    <Compile Include="Core\Champion Ports\Riven\Flowers%27 Riven\Manager\Events\Games\Mode\Burst.cs" />
+    <Compile Include="Core\Champion Ports\Riven\Flowers%27 Riven\Manager\Events\Games\Mode\Combo.cs" />
+    <Compile Include="Core\Champion Ports\Riven\Flowers%27 Riven\Manager\Events\Games\Mode\Flee.cs" />
+    <Compile Include="Core\Champion Ports\Riven\Flowers%27 Riven\Manager\Events\Games\Mode\Harass.cs" />
+    <Compile Include="Core\Champion Ports\Riven\Flowers%27 Riven\Manager\Events\Games\Mode\JungleClear.cs" />
+    <Compile Include="Core\Champion Ports\Riven\Flowers%27 Riven\Manager\Events\Games\Mode\KeepQ.cs" />
+    <Compile Include="Core\Champion Ports\Riven\Flowers%27 Riven\Manager\Events\Games\Mode\KillSteal.cs" />
+    <Compile Include="Core\Champion Ports\Riven\Flowers%27 Riven\Manager\Events\Games\Mode\LaneClear.cs" />
+    <Compile Include="Core\Champion Ports\Riven\Flowers%27 Riven\Manager\Events\Games\Mode\QADelay.cs" />
+    <Compile Include="Core\Champion Ports\Riven\Flowers%27 Riven\Manager\Events\Gapcloser\GapcloserManager.cs" />
+    <Compile Include="Core\Champion Ports\Riven\Flowers%27 Riven\Manager\Events\Interrupt\InterruptManager.cs" />
+    <Compile Include="Core\Champion Ports\Riven\Flowers%27 Riven\Manager\Events\SpellCast\SpellCastManager.cs" />
+    <Compile Include="Core\Champion Ports\Riven\Flowers%27 Riven\Manager\Menu\MenuManager.cs" />
+    <Compile Include="Core\Champion Ports\Riven\Flowers%27 Riven\Manager\Spells\SpellManager.cs" />
     <Compile Include="Core\Champion Ports\Riven\Flowers%27 Riven\myCommon\CheckStatus.cs" />
     <Compile Include="Core\Champion Ports\Riven\Flowers%27 Riven\myCommon\DamageCalculate.cs" />
     <Compile Include="Core\Champion Ports\Riven\Flowers%27 Riven\myCommon\DamageIndicator.cs" />
